Breaks out the c.a.ionsql package into c.a.ionsql.[eval,syntax,util,tools,tools.example]
The main rationale for this change is that everything today is in one flat package, and
this is the time, before we get too many dependencies to modularize our namespace
to make it easier to manage.

End users really just have to switch to syntax/eval if their embedding the parser and/or
the evaluator.

cr https://code.amazon.com/reviews/CR-106880 [no autoclose]