Adds evaluator support for UNPIVOT and AT in FROM sources.
Provides customizable delegation in BaseExprValue.
Adds the Named facet to encapsulate a name/ordinal of a value.
Adds adapters for ExprValue and Named facets.
Adds ExprValueType as a first-class type for the Evaluator.
Adds the beginning implementation of MISSING in the Evaluator.

cr https://code.amazon.com/reviews/CR-107425 [no autoclose]