IONSQL-87 Proposal to separate evaluator test data
Note: This commit is only to validate the proposal, will migrate all
existing tests and squash before submitting final CR

Idea is to change the eval and assertEval methods, suffixed temporarily
with "2", to accept a Map<String, String> that represent the bindings.
Se migrated examples on EvaluatingCompilerTest.kt

This way it's on the test implementer the decision to share data
between tests instead of forcing a global binding.

If lots of tests share the same bindings we can wrap the new version of
assertEval to always append a base binding map as syntax sugar.

cr https://code.amazon.com/reviews/CR-478291