Adds support for SQL-92 and generalized Ion SQL++ aggregation.
Adds ExprAggregator and ExprAggregatorFactory to encapsulate aggregate
functions as an accumulator.

Adds the notion of "registers" that the compiler can have in the
environment in the form of Register and RegisterBank.  This is used
for managing aggregation state, but can be used in the future for other
things such as more efficient local variables.

Refactors the threaded-code compiler such that the table driving the
compilation over the AST can be polymorphic (effectively a sort of v-table).
This allows the compiler to deal with context-sensitive Ion SQL++ constructs
such as SQL-compatible aggregation.

Note that users cannot at this time add aggregation functions, like we
can for user functions.

https://sim.amazon.com/issues/IONSQL-30

cr https://code.amazon.com/reviews/CR-518399 [no autoclose]