ION-764 Differentiate internal and external errors during eval
Adding a flag to EvaluatorException to mark it as an internal exception.
Doing this instead of creating two Exception types to make it easier to
migrate to the final impl where we'll propagate the error codes and full
error context

By external we mean errors generated due to bad user input, for example
/ by 0 and 1 + 'a'

By internal we mean errors generated due to some unexpected behavior
inside the evaluator which is potentially a bug. For example
NullPointerExceptions, InvalidStateExceptions

The internal flag should be deprecated in favor of defining and
propagating the evaluation error codes and error context. With this the
client will be in a better position to decide what should it consider as
internal and external

cr https://code.amazon.com/reviews/CR-827384