Packages the cli jar as a zip archive
Packages the cli jar into a zip archive containing:
* The jar itself
* A windows batch script
* A unix shell script
* A Readme

both scripts can be double clicked to open a terminal with the REPL.
Tested on both windows and Mac

The Readme is empty for now as a placeholder. It'll include information
regarding the tutorial when that's ready. The tutorial PDF should also
be included in the zip archive