adds REPL commands
Refactors the REPL to have a more explicit state machine so we can add
REPL specific commands.

Two commands added:
* list_commands: list all REPL commands
* add_to_global_env: adds a value to the REPL global environment. This
is useful for users that want to add new "tables" in the repl without
having to use a file.

To use it:
PartiQL> !add_to_global_env {'myTable': <<{'a': 1}, {'a': 2}>>}

will add a binding from 'myTable' to <<{'a': 1}, {'a': 2}>>

I'll add user documentation and more tests before merging. Want to get
some early feedback on this change