REPL command dump_global_env and test refactors
This commit introduces a new REPL command that dumps the current REPL
global env

It also refactors the REPL tests so they can be defined by an example of
what is the expected REPL prompt. The initial implementation of this is
using threads to minimize changes to the REPL itself but ideally we'd
implement some trace logging in the REPL that can be used to inspect
what happened. Ran each test 5k times as a sanity check against race
conditions and all passed