Issue #118, generate pom.xml and sign build artifacts.
This is being done in preparation for publishing to maven central.

To ensure the ability to publish development versions of the library to
the local maven repository for developers that do not have the GPG keys
installed, we only attempt to sign the build artifacts when we are not
building a "-SNAPSHOT" version.