Update unit tests
PR #154 introduced `SelectStarRewriter` which changes `SelectListItemStar`
into one `SelectListItemProjectAll` for each FROM source in a query.

Unfortunately, this also exposed issue #169 wherein
`SelectListItemProjectAll` (which represents `<expr>.*` in a select list)
does not behave correctly when `<expr>` is a list or bag.  That broke
unit tests for `SELECT DISTINCT *`.

This commit splits non-broken functionality of the `SELECT DISTINCT *`
unit test and `@Ignore`s new tests covering the broken functionatliy.

This commit also adds a couple of unit tests to `SelectStarRewriter`
which ensures the `DISTINCT` quantifier is being propagated to the
rewritten `ExprNode`.