Fix #121 and 314: stop treating date parts as if they are string literals
This also required adding/enhancing parser support for the special
functions extract, date_add and date_diff functions all of which
have special syntax.

This is "technically" a breaking chance as evidenced by the
changes required to the tests related to these which previously
expected these keywords to be values instead of the grammatical
elements that they actually are.  For instance, this commit
removes several tests that attempt to pass NULL and MISSING
values instead of a date part to these functions.