Inital definition of partiql_algebra and conversion from partiql_ast (#476)
This is the absolute beginning of PartiQL's experimental relational algebra which will be used to optimize PartiQL queries.  It's very early in its design so far.  For instance, many more operators other than the 4 defined so far are needed.  These will be defined later, incrementally.

This PR will merge to the `query-planner` branch where we will be also be building an experimental planner. 

- `docs/dev/RELATIONAL-ALGEBRA.md`  is intended to be the long term home for the documentation of the relational algebra.  Start there for details.
- `partiql.ion` includes a permuted PIG domain for the algebra. 
- Also included is a `astToAlgebra` function which performs conversion from the AST to the relational algebra.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.