Refactor EvaluatorTestBase To Allow Testing of Multiple Compilers (#573)
The purpose of this PR is to prepare the unit test suite to be able to test multiple compilers.

EvaluatorTestBase (ETB) class was entirely rewritten. EvaluatorTestAdapter and implementations of it are new. The other changes are (more or less) mechanical refactors that call into the new ETB functions (the bulk of the diffs in this commit). However, the test cases are almost entirely unchanged.

For ETB, there used to be about 8 protected functions (including overloads) that were for:

- "happy path" scenarios (i.e. given a query and a session, assert a particular value was returned), or...
- error scenarios (i.e. given the same, assert a particular error was thrown)

Each of these functions were slightly different. Replaces all of these functions in ETB with runEvalutorTestCase and runEvaluatorErrorTestCase which have all the capabilities of the prior functions and updated all of the old call sites to call one of these two functions. One of the challenges here is that, for the "happy path" tests, there are 4 different formats for expected values (see: ExpectedResultFormat to see how I handled this... 

Each of the new ETB functions creates an instance of EvaluatorTestCase or EvaluatorErrorTestCase which contain the data for the test case and passes it to the appropriate function on EvaluatorTestAdapter to perform the actual tests and assertions.

EvaluatorTestAdapter is an interface with a few implementations--one runs the test in multiple other adapters, one that runs the test using the AST evaluator, one that tests the legacy serializers, etc... In the near future, an implementation of EvaluatorTestAdapter will test the new query planner & physical plan compiler.

Also, the test DSL was also are modified to use a EvaluatorTestAdapter.

In the future, an implementation of EvaluatorTestAdapter could be created that emits test case data as Ion, possibly as input to a "compliance suite" of some sort. This would be a first step toward converting these unit tests to "test scripts" that could then be used to test other implementations of PartiQL.