Add CHANGELOG.md
This commit includes the changes for adding the changelog to
the project with backfilling the previous releases to `CHANGELOG.md`
along with updating `CONTRIBUTE.md` for reflecting updating the
changelog as a process change. With this change we expect to make
future releases easier and provide more visiblity on changes being
added to between releases.