Revisit FromSourceAliasVisitorTransform implementation.
The previous implementation relied on delicate ping-ponging of control between multiple instances of the host (FromSourceAliasVisitorTransform) and helper (InnerFromSourceAliasVisitorTransform) visitors.
In the amended implementation, the host visitor drives the traversal and calls a fresh instance of the helper visitor in each spot where a transformation is needed.

Both implementations create the same number of helper visitors (one per a FROM clause).
The new implementation needs only one host visitor, as opposed to a host visitor per subquery.
Absence of ping-ponging means the "stack" of visitors is at most two, as opposed to growing as the traversal goes deeper into the query, with implications for the lifetime of the involved visitor objects and, therefore, memory usage.