Model experimental `Schema` (#1016)
Adds an experimental model for PartiQL Schema; as part of working on this PR the following spec issues created as pre-requisite for finalizing the SchemaType:

- Define PartiQL Type semantics: https://github.com/partiql/partiql-spec/issues/49
- Create PartiQLSchema specification: https://github.com/partiql/partiql-docs/issues/37

With this change we can model the `SQL` and `NoSQL` schemas on collections with collection constraints. In addition, with this model, we're able to represent other schemas like nested, array, scalar, and possibly Graph schemas as collection of Graph types:

```
// Example of nested schema
<<{'a': INT, 'b': [{'x': DECIMAL, 'y': DECIMAL}]>>

// Example of Scalar schema
<<INT>>

// Example of Array schema
<<[INT, DECIMAL]>>
```

As StaticType currently does not model the Orderedness of Bag and List as
constraints, leaving the Orderedness to the BAG and LIST types themselves.

Why the experimental model is a collection?

We model the Schema (E.g. an SQL Table`) as a collection to solve the orthogonality problem that SQL has with Table (as entities that define the shape of data) which implicitly defines them as a collection.