Make EvaluatingCompiler aware of globally-bound variables.
Done by adding a GlobalsCheck class parameter -- an object that can tell whether a given variable is defined in the global environment.  For now, it is implemented as a wrapper around an EvaluationSession.
To make this addition work throughout the code base, the primary needed adjustment was in the execution order of creating a pipeline + calling compilation vs session creation -- a session now has to be created first.

Overall, this addition can be seen as a precursor to properly introducing static environment into compilation.