An initial implementation of graph query evaluation (#1074)
An end-to-end implementation of a fragment of graph pattern matching that can be unit-tested and experimented with in REPL.

The targeted language definition is what was proposed in [RFC-0025](https://github.com/partiql/partiql-docs/blob/main/RFCs/0025-graph-data-model.md) and [RFC-0033](https://github.com/partiql/partiql-docs/blob/main/RFCs/0033-graph-query.md).

The fragment of GPML patterns that is implemented can be characterized as: straight-path patterns with simple label matching and all directed/undirected edge patterns. 
This already allows one to write some interesting queries can be written and tested already.  Some examples are in the unit tests `EvaluatingCompilerGraphMatchTests`.

The evaluation approach is meant to follow the one outlined in Section 6 of the GPML paper [[ACM](https://dl.acm.org/doi/10.1145/3514221.3526057), [arxiv](https://arxiv.org/abs/2112.06217).]
In particular, the computational core of this initial implementation corresponds to Section 6.4, Computation of Path Binding, which deals with elementary "step" patterns and joining their results.  

The most interesting parts are probably the graph abstraction `Graph.kt` and some ideas for graph pattern "plans" in `MatchSpec.kt`, for structuring the matching algorithms in `GraphEngine.kt`.