A start on propagating the changes in identifier normalization into the planner pipeline.
This commit tweaks enough in AstToLogicalVisitorTransform to make tests in EvaluatingCompilerFromLetTests pass,
but there are further failures. It might be possible to fix them through similar tweaks in the subsequent transforms,
past the Logical AST.

A digression for later:

These tweaks are also fairly ugly, which might provide motivation to revise somewhat the design of transitioning to
SQL-conformant identifiers, but this might be best after the transition is complete and more polished for the EvaluatingCompiler pipeline.
The revision might amount to abandoning the idea that PartiqlAst.Id/Defnid carries the precise text of the identifier
(how it was written at a given occurrence in a query). (The motivation for that was supporting textual round-tripping better,
but that is not achievable for keywords already, so perhaps was a fool's errand to start with.)
Instead, the PartiqlAst.Id would carry just the normalized text of the identifier, without even the info whether it
originated as regular or delimited. (Note that this simplification is made possible by transitioning
to SQL-conformant identifiers.  Keeping the internal Ident class would be advisable even in this case.)

~