An Ident-based version of the Bindings class and its companions.
The new implementations are in the package org.partiql.lang.eval.binding, using the same class names as the originals.
This should allow for a gradual transition and perhaps limiting the churn in the result.

The new implementations are, so far, just a copy-and-simplify of the originals.
Further editing might be useful.