remodel err/missing nodes
The previous approach have the error/missing node to be a wrapper around the original problematic node.

Based on the discussion, we don't want to increase the public api surface.

So in this PR, we change the error/missing node to contains a message and a trace, which is a list of rex.op. We replace the original problematic node with the error/missing node, and add their children to the trace.