Drop the simple memory backed database example and its related tests.
This used to be a good example of how to implement an OLTP database with PartiQL-Kotlin, but it uses components which will be removed in the next commit which will add `UPDATE` to and fix the way we model DML in the `partiql_logical` and later domains.  Nobody has time to update this thing, and while unfortunate, its loss is not great.